ui: Be more robust about clearing the line
Some terminals (like urxvt) will allow you to "backspace" (\x08) past the
beginning of the current line, in to the previous line.  This means that
each update to the status line moves farther and farther *up* the screen.

Instead, prefix each line with "\r" to move the cursor to the beginning of
the line, and suffix each line with the "\1B[K" escape sequence to clear
from the cursor to the end of the line.

While we're at it, add comments explaining each of the escape sequences.

v2:
 - Use "\0[K" instead of overwriting with whitespace.
 - Add comments.
 - Add a CHANGELOG entry.